add sampler
