+++ src/utils/transforms.py
@@ -11,6 +11,29 @@ from random import sample
 import albumentations as A
 from albumentations.pytorch import ToTensorV2
 
+def resize_variable_size(w, h, expected_h, min_w, max_w):
+    new_w = int(expected_h * float(w) / float(h))
+    round_to = 10
+    new_w = math.ceil(new_w/round_to)*round_to
+    new_w = max(new_w, min_w)
+    new_w = min(new_w, max_w)
+
+    return new_w, expected_h
+
+class VariableResize(object):
+    def __init__(self, expected_h: int = 70, min_w: int = 100, max_w: int = 300):
+        self.expected_h = expected_h
+        self.min_w = min_w
+        self.max_w = max_w
+    
+    def __call__(self, img: Image.Image) -> np.ndarray:
+        w, h = img.size
+        img = img.convert('RGB')
+        new_w, new_h = resize_variable_size(w, h, self.expected_h, self.min_w, self.max_w)
+        img = img.resize((new_w, new_h), Image.ANTIALIAS)
+        img = np.asarray(img)
+        return img
+
 class Resize(object):
     def __init__(self, size):
         """
@@ -63,7 +86,6 @@ def resize_with_padding(img, expected_size):
 class ToTensor(object):
     def __call__(self, img: np.ndarray) -> Tensor:
         img = img.transpose(2, 0, 1)
-        img = img / 225
         return tensor(img, dtype= float32)
 
 class ToUInt8(object):
@@ -167,46 +189,6 @@ class AlbumentationsWithPadding(object):
         img = augmented['image']
         return img
 
-# the fucking imgaug lib is dogshit -> np.bool decaped
-# class VietOCRAug(object):
-#     def __init__(self) -> None:
-#         sometimes = lambda aug: iaa.Sometimes(0.3, aug)
-
-#         self.aug = iaa.Sequential(iaa.SomeOf((1, 5), 
-#             [
-#             # blur
-
-#             sometimes(iaa.OneOf([iaa.GaussianBlur(sigma=(0, 1.0)),
-#                                 iaa.MotionBlur(k=3)])),
-            
-#             # color
-#             sometimes(iaa.AddToHueAndSaturation(value=(-10, 10), per_channel=True)),
-#             sometimes(iaa.SigmoidContrast(gain=(3, 10), cutoff=(0.4, 0.6), per_channel=True)),
-#             sometimes(iaa.Invert(0.25, per_channel=0.5)),
-#             sometimes(iaa.Solarize(0.5, threshold=(32, 128))),
-#             sometimes(iaa.Dropout2d(p=0.5)),
-#             sometimes(iaa.Multiply((0.5, 1.5), per_channel=0.5)),
-#             sometimes(iaa.Add((-40, 40), per_channel=0.5)),
-
-#             sometimes(iaa.JpegCompression(compression=(5, 80))),
-            
-#             # distort
-#             sometimes(iaa.Crop(percent=(0.01, 0.05), sample_independently=True)),
-#             sometimes(iaa.PerspectiveTransform(scale=(0.01, 0.01))),
-#             sometimes(iaa.Affine(scale=(0.7, 1.3), translate_percent=(-0.1, 0.1), 
-#     #                            rotate=(-5, 5), shear=(-5, 5), 
-#                                 order=[0, 1], cval=(0, 255), 
-#                                 mode=ia.ALL)),
-#             sometimes(iaa.PiecewiseAffine(scale=(0.01, 0.01))),
-#             sometimes(iaa.OneOf([iaa.Dropout(p=(0, 0.1)),
-#                                 iaa.CoarseDropout(p=(0, 0.1), size_percent=(0.02, 0.25))])),
-
-#         ],
-#             random_order=True),
-#         random_order=True)
-
-#     def __call__(self, img: np.ndarray) -> np.ndarray:
-#         return self.aug.augment_image(img)
 
 class AlbumentationsTransform(object):
     def __init__(self, size: Tuple[int, int]):
@@ -265,3 +247,7 @@ class AlbumentationsTransform(object):
         augmented = self.aug(image=img)
         img = augmented['image']
         return img
+
+class VarialeSizeAugmenter(object):
+    def __init__(self):
+        self.aug = Compose([])
