+++ train.py
@@ -1,0 +1,14 @@
+from src.models.model import Net
+from src.models.lit_module import OCRLitModule
+from src.mdels.tokenizer import Tokenizer
+from src.data.datamodule import OCRDataModule
+from src.data.components.collator import Collator
+from src.data.components.sampler import VariableSizeSampler
+
+from torch import set_float32_matmul_precision
+from torch.optim import AdamW
+
+from lightning import Trainer
+from lightning.pytorch import loggers
+from lightning.pytorch.callbacks import ModelCheckpoint, LearningRateMonitor, EarlyStopping
+from lightning.pytorch.profilers import AdvancedProfiler
